Checkpoint moving making the producer interface be a RowWithEvent, instead of a RowMap.
Not yet done. I realize I will need to make the RowWithEvent be one of a couple different types: Update/Write/Delete, since they need to be serialized to JSON in different ways. Which likely means I will need RowWithEvent to be a abstract class. Not sure I like that. Anyway, checkpointing for now.